requirements.txt
langchain
langchain-community
google-generativeai
streamlit
langchain-google-genai
python-dotenv
chromadb
pypdf
PyPDF2
tornado
portpicker